+++ src/main/java/pl/futurecollars/invoicing/db/file/IdService.java
@@ -10,7 +10,7 @@ public class IdService {
   private final Path idFilePath;
   private final FileService fileService;
 
-  private int nextId = 1;
+  private long nextId = 1;
 
   public IdService(Path idFilePath, FileService fileService) {
     this.idFilePath = idFilePath;
@@ -29,7 +29,7 @@ public class IdService {
 
   }
 
-  public int readLastIdFromFile() {
+  public long readLastIdFromFile() {
     try {
       fileService.writeToFile(idFilePath, String.valueOf(nextId + 1));
       return nextId++;
